src/DataApp/requirements.txt
bson
paho-mqtt
influxdb-client
